# Conflicts:
#	src/octoprint/server/__init__.py
#	src/octoprint/static/js/lib/jquery/jquery.flot.time.js

+++ src/octoprint/plugins/softwareupdate/updaters/pip.py
@@ -15,12 +15,14 @@ from .. import exceptions
 logger = logging.getLogger("octoprint.plugins.softwareupdate.updaters.pip")
 console_logger = logging.getLogger("octoprint.plugins.softwareupdate.updaters.pip.console")
 
+_ALREADY_INSTALLED = "Requirement already satisfied (use --upgrade to upgrade)"
+
 _pip_callers = dict()
 _pip_version_dependency_links = pkg_resources.parse_version("1.5")
 
-def can_perform_update(target, check):
+def can_perform_update(target, check, online=True):
 	pip_caller = _get_pip_caller(command=check["pip_command"] if "pip_command" in check else None)
-	return "pip" in check and pip_caller is not None and pip_caller.available
+	return "pip" in check and pip_caller is not None and pip_caller.available and (online or check.get("offline", False))
 
 def _get_pip_caller(command=None):
 	key = command
@@ -35,10 +37,13 @@ def _get_pip_caller(command=None):
 
 	return _pip_callers[key]
 
-def perform_update(target, check, target_version, log_cb=None):
+def perform_update(target, check, target_version, log_cb=None, online=True, force=False):
 	pip_command = None
 	if "pip_command" in check:
 		pip_command = check["pip_command"]
+
+	if not online and not check.get("offline", False):
+		raise exceptions.CannotUpdateOffline()
 
 	pip_caller = _get_pip_caller(command=pip_command)
 	if pip_caller is None:
@@ -63,10 +68,10 @@ def perform_update(target, check, target_version, log_cb=None):
 		pip_caller.on_log_stdout = _log_stdout
 		pip_caller.on_log_stderr = _log_stderr
 
-	install_arg = check["pip"].format(target_version=target_version)
+	install_arg = check["pip"].format(target_version=target_version, target=target_version)
 
 	logger.debug(u"Target: %s, executing pip install %s" % (target, install_arg))
-	pip_args = ["install", check["pip"].format(target_version=target_version, target=target_version)]
+	pip_args = ["install", install_arg]
 
 	if "dependency_links" in check and check["dependency_links"]:
 		pip_args += ["--process-dependency-links"]
@@ -74,12 +79,17 @@ def perform_update(target, check, target_version, log_cb=None):
 	returncode, stdout, stderr = pip_caller.execute(*pip_args)
 	if returncode != 0:
 		raise exceptions.UpdateError("Error while executing pip install", (stdout, stderr))
+	
+	if not force and any(map(lambda x: x.strip().startswith(_ALREADY_INSTALLED) and (install_arg in x or install_arg in x.lower()), stdout)):
+		logger.debug(u"Looks like we were already installed in this version. Forcing a reinstall.")
+		force = True
 
-	logger.debug(u"Target: %s, executing pip install %s --ignore-reinstalled --force-reinstall --no-deps" % (target, install_arg))
-	pip_args += ["--ignore-installed", "--force-reinstall", "--no-deps"]
-
-	returncode, stdout, stderr = pip_caller.execute(*pip_args)
-	if returncode != 0:
-		raise exceptions.UpdateError("Error while executing pip install --force-reinstall", (stdout, stderr))
+	if force:
+		logger.debug(u"Target: %s, executing pip install %s --ignore-reinstalled --force-reinstall --no-deps" % (target, install_arg))
+		pip_args += ["--ignore-installed", "--force-reinstall", "--no-deps"]
+	
+		returncode, stdout, stderr = pip_caller.execute(*pip_args)
+		if returncode != 0:
+			raise exceptions.UpdateError("Error while executing pip install --force-reinstall", (stdout, stderr))
 
 	return "ok"
