+++ src/octoprint/server/api/files.py
@@ -29,6 +29,8 @@ import threading
 _file_cache = dict()
 _file_cache_mutex = threading.RLock()
 
+_DATA_FORMAT_VERSION = "v2"
+
 def _clear_file_cache():
 	with _file_cache_mutex:
 		_file_cache.clear()
@@ -78,6 +80,8 @@ def _create_etag(path, filter, recursive, lm=None):
 	if path.endswith("/files") or path.endswith("/files/sdcard"):
 		# include sd data in etag
 		hash.update(repr(sorted(printer.get_sd_files(), key=lambda x: x[0])))
+
+	hash.update(_DATA_FORMAT_VERSION) # increment version if we change the API format
 
 	return hash.hexdigest()
 
@@ -148,6 +152,7 @@ def _getFileList(origin, path=None, filter=None, recursive=False, allow_from_cac
 				file = {
 					"type": "machinecode",
 					"name": sdFile,
+					"display": sdFile,
 					"path": sdFile,
 					"origin": FileDestinations.SDCARD,
 					"refs": {
@@ -287,8 +292,11 @@ def uploadGcodeFile(target):
 		# determine future filename of file to be uploaded, abort if it can't be uploaded
 		try:
 			# FileDestinations.LOCAL = should normally be target, but can't because SDCard handling isn't implemented yet
-			futurePath, futureFilename = fileManager.sanitize(FileDestinations.LOCAL, upload.filename)
+			canonPath, canonFilename = fileManager.canonicalize(FileDestinations.LOCAL, upload.filename)
+			futurePath = fileManager.sanitize_path(FileDestinations.LOCAL, canonPath)
+			futureFilename = fileManager.sanitize_name(FileDestinations.LOCAL, canonFilename)
 		except:
+			canonFilename = None
 			futurePath = None
 			futureFilename = None
 
@@ -316,7 +324,7 @@ def uploadGcodeFile(target):
 			Depending on the file's destination triggers either streaming to SD card or directly calls selectAndOrPrint.
 			"""
 
-			if destination == FileDestinations.SDCARD and octoprint.filemanager.valid_file_type(filename, "gcode"):
+			if destination == FileDestinations.SDCARD and octoprint.filemanager.valid_file_type(filename, "machinecode"):
 				return filename, printer.add_sd_file(filename, absFilename, selectAndOrPrint)
 			else:
 				selectAndOrPrint(filename, absFilename, destination)
@@ -335,7 +343,9 @@ def uploadGcodeFile(target):
 				printer.select_file(absFilename, destination == FileDestinations.SDCARD, printAfterSelect)
 
 		try:
-			added_file = fileManager.add_file(FileDestinations.LOCAL, futureFullPathInStorage, upload, allow_overwrite=True)
+			added_file = fileManager.add_file(FileDestinations.LOCAL, futureFullPathInStorage, upload,
+			                                  allow_overwrite=True,
+			                                  display=canonFilename)
 		except octoprint.filemanager.storage.StorageError as e:
 			if e.code == octoprint.filemanager.storage.StorageError.INVALID_FILE:
 				return make_response("Could not upload the file \"{}\", invalid type".format(upload.filename), 400)
@@ -401,7 +411,9 @@ def uploadGcodeFile(target):
 		if not target in [FileDestinations.LOCAL]:
 			return make_response("Unknown target: %s" % target, 400)
 
-		futurePath, futureName = fileManager.sanitize(target, foldername)
+		canonPath, canonName = fileManager.canonicalize(target, foldername)
+		futurePath = fileManager.sanitize_path(target, canonPath)
+		futureName = fileManager.sanitize_name(target, canonName)
 		if not futureName or not futurePath:
 			return make_response("Can't create a folder with an empty name", 400)
 
@@ -414,7 +426,7 @@ def uploadGcodeFile(target):
 			return make_response("Can't create a folder named %s, please try another name" % futureName, 409)
 
 		try:
-			added_folder = fileManager.add_folder(target, futureFullPath)
+			added_folder = fileManager.add_folder(target, futureFullPath, display=canonName)
 		except octoprint.filemanager.storage.StorageError as e:
 			if e.code == octoprint.filemanager.storage.StorageError.INVALID_DIRECTORY:
 				return make_response("Could not create folder {}, invalid directory".format(futureName))
@@ -464,6 +476,7 @@ def gcodeFileCommand(filename, target):
 	valid_commands = {
 		"select": [],
 		"slice": [],
+		"analyse": [],
 		"copy": ["destination"],
 		"move": ["destination"]
 	}
@@ -539,6 +552,14 @@ def gcodeFileCommand(filename, target):
 			if path:
 				full_path = fileManager.join_path(target, path, destination)
 
+		canon_path, canon_name = fileManager.canonicalize(target, full_path)
+		sanitized_name = fileManager.sanitize_name(target, canon_name)
+
+		if canon_path:
+			full_path = fileManager.join_path(target, canon_path, sanitized_name)
+		else:
+			full_path = sanitized_name
+
 		# prohibit overwriting the file that is currently being printed
 		currentOrigin, currentFilename = _getCurrentFile()
 		if currentFilename == full_path and currentOrigin == target and (printer.is_printing() or printer.is_paused()):
@@ -595,6 +616,7 @@ def gcodeFileCommand(filename, target):
 			                  printer_profile_id=printerProfile,
 			                  position=position,
 			                  overrides=overrides,
+			                  display=canon_name,
 			                  callback=slicing_done,
 			                  callback_args=(target, full_path, select_after_slicing, print_after_slicing))
 		except octoprint.slicing.UnknownProfile:
@@ -603,8 +625,9 @@ def gcodeFileCommand(filename, target):
 		files = {}
 		location = url_for(".readGcodeFile", target=target, filename=full_path, _external=True)
 		result = {
-			"name": destination,
+			"name": sanitized_name,
 			"path": full_path,
+			"display": canon_name,
 			"origin": FileDestinations.LOCAL,
 			"refs": {
 				"resource": location,
@@ -616,6 +639,17 @@ def gcodeFileCommand(filename, target):
 		r.headers["Location"] = location
 		return r
 
+	elif command == "analyse":
+		if not _verifyFileExists(target, filename):
+			return make_response("File not found on '%s': %s" % (target, filename), 404)
+
+		printer_profile = None
+		if "printerProfile" in data and data["printerProfile"]:
+			printer_profile = data["printerProfile"]
+
+		if not fileManager.analyse(target, filename, printer_profile_id=printer_profile):
+			return make_response("No analysis possible for {} on {}".format(filename, target), 400)
+
 	elif command == "copy" or command == "move":
 		# Copy and move are only possible on local storage
 		if not target in [FileDestinations.LOCAL]:
@@ -625,14 +659,15 @@ def gcodeFileCommand(filename, target):
 			return make_response("File or folder not found on {}: {}".format(target, filename), 404)
 
 		path, name = fileManager.split_path(target, filename)
-		destination = data["destination"]
-		if _verifyFolderExists(target, destination):
-			# destination is an existing folder, we'll assume we are supposed to move filename to this
-			# folder under the same name
-			destination = fileManager.join_path(target, destination, name)
 
-		if _verifyFileExists(target, destination) or _verifyFolderExists(target, destination):
-			return make_response("File or folder does already exist on {}: {}".format(target, destination), 409)
+		destination = data["destination"]
+		dst_path, dst_name = fileManager.split_path(target, destination)
+		sanitized_destination = fileManager.join_path(target, dst_path, fileManager.sanitize_name(target, dst_name))
+
+		if _verifyFolderExists(target, destination) and sanitized_destination != filename:
+			# destination is an existing folder and not ourselves (= display rename), we'll assume we are supposed
+			# to move filename to this folder under the same name
+			destination = fileManager.join_path(target, destination, name)
 
 		is_file = fileManager.file_exists(target, filename)
 		is_folder = fileManager.folder_exists(target, filename)
@@ -641,13 +676,23 @@ def gcodeFileCommand(filename, target):
 			return make_response("{} on {} is neither file or folder, can't {}".format(filename, target, command), 400)
 
 		if command == "copy":
+			# destination already there? error...
+			if _verifyFileExists(target, destination) or _verifyFolderExists(target, destination):
+				return make_response("File or folder does already exist on {}: {}".format(target, destination), 409)
+
 			if is_file:
 				fileManager.copy_file(target, filename, destination)
 			else:
 				fileManager.copy_folder(target, filename, destination)
+
 		elif command == "move":
 			if _isBusy(target, filename):
 				return make_response("Trying to move a file or folder that is currently in use: {}".format(filename), 409)
+
+			# destination already there AND not ourselves (= display rename)? error...
+			if (_verifyFileExists(target, destination) or _verifyFolderExists(target, destination)) \
+					and sanitized_destination != filename:
+				return make_response("File or folder does already exist on {}: {}".format(target, destination), 409)
 
 			# deselect the file if it's currently selected
 			currentOrigin, currentFilename = _getCurrentFile()
